Transform work: Dexing, res shrinking, API.
1. Dexing
In order to simplify dex, pre-dexing (really just per-scope
dexing), multi-dexing (both legacy and native) and dex merging,
everything is moved to a single transform.

This transform either runs a single dex step (no pre-dexing),
or first runs dexing per scope, and then a dex merge.
If multi-dex is enabled:
- in Legacy mode, the merge is given a list of classes to put
  in the main dex file.
- in native mode (which enforces pre-dexing), the merge doesn't
  happen and only per-scope dexing happens directly in the output
  folder (rather than in an intermediate folder)

This means that at the end of the setup, the only streams
left in the TransformManager are RESOURCES and DEX ones.

2. Resource Shrinker.
The impact of #1 on the available streams at the end of the
setup means that res shrinking needs to move to a transform,
so that it can have access to the intermediate classes streams.
Since it needs to read the intermediate class files, but
doesn't actually output anything in the streams, it must
be setup as a NO_OP transform.

However shrinking is per variant output, not per variant, so
some changes are introduced:
- BaseScope is now the base of VariantScope and VariantOutputScope.
  This is used by the TransformManager to just handle a generic scope
  rather than have separate code for the 2 cases.
- VariantOutputData holds the TransformTask matching the actual
  transform doing the shrink.

3. Transform API
This change also includes the beginning of a public
tranform API.
- Allows to register transforms globally, on the 'android'
  extension
- These transforms apply to *all* variants. No way to make
  them per variant yet (later commit)
- No way to control the order (later commit)

The JarJarTest which relied on the ability to insert
a JarJar step in the build now relies on this (successfully.)

4. Dex task in Variant API
Since the dex step is handled by a transform we do not
want to give access to the transform task through the variant
API.
In most cases the need is handled by the Transform API anyway.

Currently missing: addtionalParams send to dex are missing. They
need to be added either globally (dexOptions) or per-variant
(later commit)

Change-Id: Ia1a1273f6943d9e5fa105ebf4e389e8f4e05576d