79325: Automatic resource shrinking causes app to crash
Apps built with resource shrinking can crash on a few specific devices
where customizations to Android are (for various reasons) crawling
through the reachable resources in the app and attempting to load
them.

This works around that behavior by rather than *removing* these unused
resources, replacing them with the smallest valid files that it can --
for .png files a 67 byte 1 pixel PNG file, for .9.png files a 73 byte
valid nine patch, and for .xml resources a minimal one element binary
XML file.

These are obviously larger than removing the files completely, but
should in all cases be smaller than the unused resource, and for
bitmaps in particular should typically be a large savings.

When aapt supports stable id's we can regenerate the ap_ file with the
unused resource completely removed (such that they are not referenced
in the .arsc file), and at that point we can remove the dummy files
too. For that reason, this behavior is controlled with a (constant
flag) rather than unconditionally changing the code and unit tests to
always behave this way.

Change-Id: I72137ac112c1d034ea1a8885b1f3dc7d7db8dd8c