Fix issue where optional libs are not passed to proguard.
If a dependency of a project uses an optional library (like
the http library that's now optional), then Proguard must
see this library or it'll break.

This fixes makes it so that all optional libraries are
always passed to proguard whether or not the module
that's built requires the libraries.

https://code.google.com/p/android/issues/detail?id=184883

Change-Id: Ia392ff16d15f2b495ab19f68edb756c666700fae