More transform work.
Fix incremental mode for dex merging.
We cannot rely on what gets re-dexed to compute the
list of input for the merge. All previous dex files need
to be included. New code/New test for this.
Plus fixed the ApkSubject to handle multi-dex.

Fixed multi-dex for test projects.
Automtically setup multi-dex library and test runner
for legacy multi-dex in test project. This ensures that
at least some classes are retained in the main classes
file (the test runner) preventing cases of empty main
classe file (which proguard would break on).
Of course, it's also good for the user to not have to
manually setup this.

Hack-fix for the test of library project. Because of
dependency problem, the local jars of the library
are added twice to the test artifact. Until we
can fix this, restrict this scope from the
computation when doing library project.
Downside is that local jars only of the test
component will be missing.
We'll fix this better soon.

Change-Id: Ie0e41629b28f38ab0a546e28d35277e02bd04faa