Transform is now an abstract class.
This allows future proofing when we can
add new methods with default behavior for
older implementations that would not implement
the new methods.

The {Type}Transform interfaces stay as interfaces
since they are meant to only provide a single method.
Also, that way they do not extend Transform anymore,
providing more flexibility when dealing with a custom
type hierarchy that starts at Transform.

Change-Id: I5ebea91756a0c3d2d097774aed3618187dc168a7