first version of the InstantRun verifier.
added a no-op transform that gets notified of all new classes
compiled and will make a copy for the next iteration.

at each iterarion, the verifier compares the copy with the new
class version and check that the changes are compatible with
the InstantRun capabilities. The first version is doing very
little, mostly to hook things up at this point.

when some class change is incompatible, the reload.dex is
not produced which is the indication for the FDR that a
restart is necessary.

Change-Id: Id2f13e05dc8cf295ba41c2e21366c4d925753dec