Fix the navigation drawer to go on top of the status bar.
This is a fix for the issue: 187516
The FreeMarker code caches source files of the same name.
The cache does not know about folders. We had 2 files in 2 different
folders which caused FreeMarker to instantiate the wrong file.
This is a quick fix. A better fix would be to let Freemarker know
about folder names. We will make that fix for 1.5.

Change-Id: I7c92e736ee860e14324ddc0a4c721a223c328545