Convert MULTI_FOLDER to SINGLE_FOLDER for transforms.
When calling a transform, convert MULTI_FOLDER to
a list of SINGLE_FOLDER with their own Transform
Input/Output, and their own list of changed files.

This allows implementation of Transforms to only
care about single folders and jars.

It also allows handling incremental work at the folder
level, instead of the stream level which was difficult
with MULTI_FOLDER.

Note: We still need to figure out removal of
streams.

Change-Id: Iebf05a795ef47a7067b59f675f3859524ace9d8f