Revamp the way we handle native libraries.
Using the same mechanism as the way we now handle
java resources, there is first a task to merge
all the libraries coming from the jniLibs folder.
This is done using the same task that merges the
Assets (which gets renamed).

Then the MergeJavaRes Transform gets applied
but with different filters to only take into account the
so files. This merges the folders, the generated so
and the ones coming from the dependencies.

Change-Id: I50d7f28593958aaebe8782585594a2838b926cf7