Add file content lookup abstraction to the manifest merger
Clients of the manifest merger can now provide a custom method to read
the contents of a File. This allows the manifest merger to

(1) Operate on the live, unsaved contents of documents in the IDE
    editor, and
(2) Merge files in memory only rather than creating and cleaning up
    temporary disk files as is done today (e.g. on project creation).

The essential part of the IDE is that the manifest merger no longer
reads files directly; instead, it consults a FileStreamProvider to map
a File to its content. This turns the File into a pure handle to
reference files. For scenario (1) above, the IDE provides a file
stream handler which maps a java.io.File back to an IDE PSI file,
which is then mapped to an editor Document which we then read
from. And for scenario (2) we simply create fake non-existent File
handles (which are not required to correspond to actual existing
files) which is mapped back to the in memory template files on the
fly.

(It's unfortunate that the API is using InputStreams instead of
Readers since clients such as the IDE may have additional knowledge
about the encoding of a file; however, since this interface lives
inside the manifest merger it is tied to XML files where encodings are
either specified within the XML file or assumed to be UTF-8 files (and
the XML APIs are using the InputStreams directly). If we want to
generalize this outside of the manifest merger in the future we should
switch to a Reader interface.)

Change-Id: I272d5281a3fe18492a1e2090a2a9186c484805c8