Use RecyclerView in List Fragment template.
The template used a ListFragment or a GridView. Both of
these are now replaced with a RecyclerView.
The automatic switch from List to Grid has been removed,
instead the number of columns are specified by the user.
Also: Remove compiler warning by updating onAttach from the
deprecated version to the version taking a Context argument.

Change-Id: I70506414e05e2c18f067798afebe02fb04dc6fe2