added annotation to suppress instant run for a package, class or method.
When code is speed sensitive, or is generating a runtime error when run
within the InstantRun framework, add a class level annotation that can be
used to annotate a package, a class or a method to disable InstantRun
for the annotated element.

Some more changes will be necessary on the verifier side when a
disabled method, class or package is changed, we should generate a
restart event.

Change-Id: I5fa91dfd2bd3b7ece59ca156fb688b494a40a746