Dex and pre-dex in-process.
This uses a pool of class loaders that can be used for dexing. Since dx
code is not thread-safe, we have to isolate instances inside class
loaders.

Class loaders are kept in a global static cache, indexed by build tools
version. Although the cache itself is kept for the lifetime of the
daemon, classloaders are discarded (and GC'd) N hours after last being
used.

The classloaders are used in the MRU order and initialized lazily, which
means that smaller projects will only initialize and re-use the minimum
number of class loaders needed (which should contain JITed code).

There is a fixed-size pool of dexers for every version of build tools
and worker threads block while waiting for a free dexer to be available
- this limits the number of files dexed in parallel. Every dexer may
spawn more threads, according to the threadCount property of DexOptions
(by default every file is dexed with 4 threads).

Change-Id: I9af451508c4f644f5f5e5a7a41b29cdd00121a0a