Generic SyncIssue for error w/o quick fixes.
The goal of having specific type ID for the SyncIssue
is to have the IDE special case them either in display,
or better, by offering quick fixes.

But there are a bunch of errors where quick fixes
don't make sense, and we can just use a generic error.

This change remove TYPE_NONE and makes it TYPE_GENERIC
and remove some recent additions which wouldn't have
quick fixes (nothing to fix really), and move them
to the generic error type.

Change-Id: I3929bbc3917f5127b57de642e21dddc00816cf93