Adds a flag to mark when a patch becomes obsolete.
For debugging purposes we want to identify if we are running obsolete code.
The other alternative, and in theory more correct, was to add a variable with
the type of the original class, and check if the $change variable was an instance
of the current class, if not then this one is obsolete. However that would
have meant longer debugger interactions.

Change-Id: If20ebaeb9be359799bbe8e7cfa0a02a1070ea0dd