Adding android-sdk specific repository logic.
This change adds information specific to the android sdk to the core
repository. This includes:
- all the standard android sdk package types
- a way to load existing local repositories and convert them to the new
  style.
- The schemas for the android remote source lists
- Initialization and handling of the RepoManager with the android info

This also includes refactoring of License (the one class that's now in
repository and also used in other contexts) into repository.
Finally, there is some small refactoring of the old sdk code that's
left over from before, but is still nice.

Change-Id: Ic7c0c9f4410131e77e5237694dc1897f3dfecdb7