Revert "Add help for template "package name" field"
The help message wasn't harmful per se, but actually it turns out it doesn't normally show up because of an issue with the Swing component that represents it, so might as well not clutter up the template for now. Also, there are lots of templates that are missing this help tooltip, so if we fix it here, we should fix it everywhere.

This reverts commit cfc298933edc71d5de014a9ecc26252b2f0e3c1c.

Change-Id: Ib7b03a6b0a494a8db026190cff1e13e1004cf62c
(cherry picked from commit 70eea9fc52cd23bfb1a100b947a9678393c2a1ed)