Fix jacoco version logic.
If the android gradle plugin is not on the classpath of the root project
then it would fail to find it, and try to create a project dependency to
the jacoco agent with a version of 'null'.

Now it looks through the parent projects until we find the dependency.

Change-Id: I03256d4fea5bd7397e2dfa500c8604146489898c
(cherry picked from commit 2ac870c9a0b22ffdd0b8f0c54a334f03d81fd2c3)