Adding android target support to new sdk
Added two new concepts: AndroidTargetManager and SystemImageManager,
both of which return all objects of the corresponding types.
Both (internally) have implementations that use both the old and new
sdk APIs, to facilitate the transition.

Also enhanced InheritancePlugin to generate a method to validate
string values with respect to pattern restristions on fields.
Change-Id: I00f2f0c43299241bfd5c997c324108af28228b9d