Make MonitorThread responsible for consuming JdwpPackets
MonitorThread sits between a Client and the debugger and monitors packets from both directions.
Currently, MonitorThread reads in a JdwpPacket, but it leaves the target of the packet to
actually "consume" the packet (which releases the packet's data in the buffer).

This CL restructures this so that MonitorThread reads a packet, dispatches it appropriately,
and then consumes the packet as well.

Note that even though this CL brings the packet read/dispatch/consume cycle to one place,
there is still follow up work needed to move it to the correct one. The owners of the buffers
(Client and Debugger) should be the ones administering it, as opposed to exposing their
internal state to MonitorThread.

Change-Id: I299a6574a6c92bf4e885b91f6ab1f4b23a5084ca