Refactors JDWP packet management to allow extensions
This is the first step at separating the pure JDWP logic
from the DDM packets. Much of the logic of package interception
is now generic for any extension to use and most of the DDM
specific modifications have been moved to DdmJdwpExtension.

In particular, this CL introduces a few abstractions:
 - A JdwpAgent is an endpoint for JDWP packets. When a packet
   is being received by an agent, it is first processed through
   a list of interceptors. If none of the interceptors are interested in
   the packet, then it is forwarded on to the target agent.
 - A JdwpExtension can register interceptors with each client and its
   Debugger when they comes online. The DDM packets are
   handled via the DdmJdwpExtension.

This CL should not have any functional effects at all.

Change-Id: I386e9d8ad40198f54de8c930cfcc04fa1f93d3a8