Fix ApiDetector to handle casts to java.lang.Object
The API database omits java.lang.Object since every single class can
be cast to it; however, this wasn't handled right for the cast
check. However, this was masked by a unit test which was referring to
an old compilation target (compileSdkVersion 3) which I had installed
on my machine, so the test passed locally, but it was failing on the
server. I've now changed the test to use something recent (23), which
also meant updating the test to be compileable (since it referred to
some no-longer existing Manifest.permission constants.)

Change-Id: I643dffd7b180058b347ecaf3153a75a9560056bc