Fix Category sorting (used by IDE inspection registration code)
Also tweak bidirectional text category to drop the hyphen, and
to be a subcategory of i18n issues.

Also tweak issue ID for GradleDetector check for plugin version
to make it have the same prefix as the other gradle checks.

Also update the "has quickfix" list for issues in HTML Reports.

Change-Id: I285a2bcae8ca08cdc8de94836adb87d7f157e22c