Make export info task incremental
This CL defines DataBindingExportInfoTask's input and outputs to
make it incremental.

This is the task data binding uses to invoke our annotation
processor. It used to run every time which was an issue if no
java/res changes are made.

Now it only runs if sources changed and only invalidates the
JavaCompile task if necessary.

Bug: 22516688

Change-Id: I193ee234e78bf47949072eba53bcf9a51214222b