187596 and 186098: Improve resource shrinker
This CL fixes a couple of problems with the resource shrinker.

First, it was not correctly skipping field referenced to resources
when it encountered an obfuscated library class (such as appcompat's
own R$string file). This meant that many appcompat resources were
incorrectly marked as used by the app.

186098: Resource shrinking does not shrink resources that referenced
        from obfuscated **.R$* classes.

Second, when the resource shrinker sees a call to
Resources#getIdentifier (which is a method which lets you look up
resources by dynamically computed resource names), it reverts to a
safe mode where it plays it safe and for example marks any resource
whose name is a match for any constant in the string pool as possibly
being used.

Unfortunately, appcompat, used in nearly all apps created from Android
Studio, has a couple of usages of Resources#getIdentifier, which means
that nearly all apps end up in safe mode, where many resources are
incorrectly kept just to be on the safe side. However, it turns out
that these two usages are benign and not typically used to look up
resources in a reflective way.

Therefore, the resource shrinker now keeps track of the obfuscated
names of the classes for these two resource lookup calls, and when it
comes across a Resources#getIdentifier() call it ignores it if it's
one of these "benign" usages.

Fixes 187596: Resource shrinking doesn't work as expected

Third, while the resource shrinker recently started replacing unused
resources with tiny placeholders instead of deleting them, it now
deletes rather than add a placeholder for the file res/raw/keep.xml.

Finally, the resource shrinker now checks to make sure there was a
net positive reduction in the resource size, and if not, it just
uses the original AAPT resource file.