Improve unused resource detector
This CL extracts and improves the resource usage analysis from the
resource shrinker such that it can be shared between the unused
resource detector in lint, and the build time resource shrinker.

This makes the unused resource detector better in the following ways:

(1) It now correctly treates resources as unused even if they are
    referenced from other resources, as long as those resources are in
    turn also recursively unused. (This fixes issue #22872).

(2) It now obeys the tools:keep and tools:discard attributes used to
    direct the resource shrinker. (Conversely, the resource shrinker
    now also obeys tools:ignore="UnusedResources").

(3) It now considers *inactive* source providers. This means that if
    is referenced but only from for example a product flavor or build
    type that is not the currently active variant in the IDE, lint
    will now discover that the resource is used after all and will not
    list it as unused.
    Fixes 60864: Unused Resources should look at all variants

(4) It now analyzes raw files, and in particular, for HTML, CSS and
    JavaScript files it can find and keep resources referenced as
    style, script and image URLs.

(5) It correctly handles @+id attribute declarations such that these
    are taken as references only if not specified on the android:id
    attribute.

(6) It exposes hooks such that IDE now offers both quickfixes for
    individual unused resources as well as for a full remove unused
    resource refactoring operation.

(7) Suppressing a resource for one of the declarations will whitelist
    all the other declaration locations for the same resource
    (this fixes issue #40647).

(8) Lint now checks static import statements and resolves variable
    references if necessary to correctly mark statically referenced
    fields as used (this fixes issue #40293)

(9) Lint's resource model is a bit more comprehensive now; this fixes
    some other long standing lint issues such as issue #25327.

TODO: For unknown reasons, having the ResourceUsageAnalyzer reference
lint's ResourceUsageModel doesn't work; it throws a
ClassNotFoundException - even though the dependencies appear to work
(and unit test execution etc works.) Therefore, for now the
ResourceUsageModel class is duplicated and appears in each place until
this can be	tracked	down.

Change-Id: I2f0013836714ba63554b0742ea20ce3dfe390415